Replace double type with time in config module.
The C code does not use doubles to represent seconds like the Perl code did so time can be represented as an integer which reduces the number of data types that config has to understand.

Also remove Variant doubles since they are no longer used.

Note that not all double code was removed since we still need to display times to the user in seconds and it is possible for the times to be fractional. In the future this will likely be simplified by storing the original user input and using that value when the time needs to be displayed.

+++ src/config/define.h
@@ -42,8 +42,8 @@ bool cfgDefOptionAllowListValueValid(ConfigCommand commandId, ConfigOption optio
 
 // Allow range
 bool cfgDefOptionAllowRange(ConfigCommand commandId, ConfigOption optionId);
-double cfgDefOptionAllowRangeMax(ConfigCommand commandId, ConfigOption optionId);
-double cfgDefOptionAllowRangeMin(ConfigCommand commandId, ConfigOption optionId);
+int64_t cfgDefOptionAllowRangeMax(ConfigCommand commandId, ConfigOption optionId);
+int64_t cfgDefOptionAllowRangeMin(ConfigCommand commandId, ConfigOption optionId);
 
 // Default value for the option
 const char *cfgDefOptionDefault(ConfigCommand commandId, ConfigOption optionId);
